backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.14
django-cors-headers>=4.3
python-dotenv>=1.0
dj-database-url>=2.1
gunicorn>=22.0
psycopg2-binary>=2.9
whitenoise>=6.6
Pillow>=10.0
